requirements.windows.cuda12_4.txt
#! Python3.9

# numpy>=1.23.0,<=2.1.1         # Installing NumPy, a package for scientific computing
# opencv-python>=4.6.0            # Installing OpenCV, the Open source Computer Vision library
# pillow>=7.1.2                   # Installing Pillow, a Python Image Library
# pyyaml>=5.3.1                   # Installing PyYAML, a library for reading configuration files
# scipy>=1.4.1                    # Installing SciPy, a library for mathematics, science, and engineering
# tqdm>=4.64.0                    # Installing TDQM, the Fast, Extensible Progress Meter

# For CUDA 12.4
--index-url https://download.pytorch.org/whl/cu124
torch        # Installing PyTorch, an open source machine learning framework
--index-url https://download.pytorch.org/whl/cu124
torchvision     # Installing TorchVision, for working with computer vision models

# Specific version because it's known to work with our code
ultralytics==8.3.85             # Installing Ultralytics package for YOLO models

# ONNX Runtime with CUDA support
onnxruntime-gpu==1.21.0         # Installing ONNX Runtime with GPU acceleration

CodeProject-AI-SDK              # Installing the CodeProject.AI SDK

# last line empty.